+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.5.0
+
+- Mod comptibility with v81 game version
+- Added Feiopar counterplay
+
 ## 1.4.1
 
 - Fixed clients not being able to place items on a Jester if it spawned inside while they were outside (or vice versa)